refactor: extract query methods to postgres_registry_query.go to stay under 800-line limit
